+++ CHANGELOG.md
@@ -6,7 +6,9 @@ This is the changelog for Cldr_units v3.9.1 released on November ____, 2021.  Fo
 
 ### Bug Fixes
 
+* Support parsing units with multiple "per" clauses like
 * Fix canonical unit name for currency units. This also fixes unit math with currency units.
+
 
 ## Cldr_Units v3.9.0